Add argument to delete URL

+++ jumpcutter.py
@@ -64,6 +64,7 @@ parser.add_argument('--frame_margin', type=float, default=1, help="some silent f
 parser.add_argument('--sample_rate', type=float, default=44100, help="sample rate of the input and output videos")
 parser.add_argument('--frame_rate', type=float, default=30, help="frame rate of the input and output videos. optional... I try to find it out myself, but it doesn't always work.")
 parser.add_argument('--frame_quality', type=int, default=3, help="quality of frames to be extracted from input video. 1 is highest, 31 is lowest, 3 is the default.")
+parser.add_argument('--url_delete', type=str, default="true", help="deletes the URL file after the jumpcutter is done REQUIRES THE URL ARGUMENT (optional. if not included, it will default to deleting the URL.), use true or false")
 
 args = parser.parse_args()
 
@@ -74,6 +75,7 @@ SAMPLE_RATE = args.sample_rate
 SILENT_THRESHOLD = args.silent_threshold
 FRAME_SPREADAGE = args.frame_margin
 NEW_SPEED = [args.silent_speed, args.sounded_speed]
+URLDelete = args.url_delete
 if args.url != None:
     INPUT_FILE = downloadFile(args.url)
 else:
@@ -87,6 +89,10 @@ if len(args.output_file) >= 1:
     OUTPUT_FILE = args.output_file
 else:
     OUTPUT_FILE = inputToOutputFilename(INPUT_FILE)
+
+
+    
+
 
 TEMP_FOLDER = "TEMP"
 AUDIO_FADE_ENVELOPE_SIZE = 400 # smooth out transitiion's audio by quickly fading in/out (arbitrary magic number whatever)
@@ -202,13 +208,14 @@ command = "ffmpeg -framerate "+str(frameRate)+" -i "+TEMP_FOLDER+"/newFrame%06d.
 subprocess.call(command, shell=True)
 
 deletePath(TEMP_FOLDER)
-try:
-    newname
-except NameError:
-    print("")
+
+if URLDelete == "true":
+    try:
+        newname
+    except NameError:
+        assert URL != None, "why u use this argument but no url, that dum"
+    else:
+        if(newname != None):
+            os.remove('%s' % newname)
 else:
-    if(newname != None):
-        os.remove('%s' % newname)
-    
-
-
+    print("Ok.")
